requirements.test.txt
pytest
pytest-cov>=2.9.0
pytest-homeassistant
